+++ src/main/java/com/example/ecommerceproject/service/BuyerService.java
@@ -1,0 +1,5 @@
+package com.example.ecommerceproject.service;
+
+public interface BuyerService {
+
+}
