+++ src/main/java/com/example/ecommerceproject/service/SellerService.java
@@ -1,0 +1,7 @@
+package com.example.ecommerceproject.service;
+
+import com.example.ecommerceproject.domain.dto.ItemFormDto;
+
+public interface SellerService {
+  public String addItem(ItemFormDto itemFormDto);
+}
