+++ src/main/java/com/example/ecommerceproject/service/impl/BuyerServiceImpl.java
@@ -1,0 +1,5 @@
+package com.example.ecommerceproject.service.impl;
+
+public class BuyerServiceImpl {
+
+}
