+++ src/main/java/com/example/ecommerceproject/service/impl/SellerServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.ecommerceproject.service;
+package com.example.ecommerceproject.service.impl;
 
 import com.example.ecommerceproject.domain.dto.ItemFormDto;
 import com.example.ecommerceproject.domain.model.Item;
@@ -8,7 +8,7 @@ import com.example.ecommerceproject.exception.BusinessException;
 import com.example.ecommerceproject.exception.ErrorCode;
 import com.example.ecommerceproject.repository.ItemRepository;
 import com.example.ecommerceproject.repository.MemberRepository;
-import com.example.ecommerceproject.repository.StockRepository;
+import com.example.ecommerceproject.service.SellerService;
 import com.example.ecommerceproject.util.ValidUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,10 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class ItemService {
+public class SellerServiceImpl implements SellerService {
 
   private final MemberRepository memberRepository;
   private final ItemRepository itemRepository;
+
 
   // @Transactional 사용 이유
   // 만약 트랜잭션을 사용하지 않았다면, 각각의 DB 작업(INSERT, UPDATE 등등)은 별개의 트랜잭션으로 취급됨
@@ -28,10 +29,9 @@ public class ItemService {
   // addItem에서는 새로운 상품을 등록하고, 이에 대한 재고를 등록함
   // 만약 상품 등록은 성공했지만, 재고 등록에서 에러가 발생한다면???, 상품 등록은 이미 커밋되었기에 롤백이 불가능함
   // 이를 방지하기 위해 사용, 동시성 문제 해결하는 것 X
+  @Override
   @Transactional
   public String addItem(ItemFormDto itemFormDto) {
-
-    // 사용자 조회및 권한 체크
     Member member = memberRepository.findById(itemFormDto.getSellerId())
         .orElseThrow(() -> new BusinessException(ErrorCode.SELLER_NOT_FOUND));
